Organizing release package: move imageSelector and serviceSelector to separate files
